The PSA implementation is production-quality
This has been the case for a while, but we forgot to update the readme.

Don't prominently label it a "reference" implementation. That implies that
it's a complete implementation, but it isn't: we do not intend to implement
every mechanism that the PSA specification has an encoding for. That also
tends to imply that it's for demonstration purposes and not ready for
production, but Mbed TLS is intended to be used in production.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>