Try to show error names in TEST_EQUAL
In TEST_EQUAL, use heuristics to guess the category (type) of the values
that are compared. Based on these heuristics, and on the C code generated by
generate_value_names.py, try to display the symbolic name of a value.

This is currently implemented for psa_status_t values (PSA_ERROR_xxx) and
for Mbed TLS error codes (MBEDTLS_ERR_xxx). For Mbed TLS error codes, also
try to break into high-level and low-level error codes.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>