Fix failure in some configs by being more precise
This test was failing in test_psa_crypto_config_accel_cipher_aead
because there are hits to the key management driver from CTR-DRBG.

This is fixed by counting hits not just for the whole key management
area, but for specific entry point, making the test more specific and
more robust.

Also fix the location of psa_crypto_init() relative to forcing errors.

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>