Fix dlopen build in the full config
Fix the build of the dlopen program when MBEDTLS_PLATFORM_PRINTF_ALT or
MBEDTLS_PLATFORM_EXIT_ALT is enabled. In these configurations,
mbedtls_printf and mbedtls_exit are linked from libmbedcrypto, so they can't
be called directly from the program code. Use the standard functions
directly. This is a test program that isn't meant to be portable to the kind
of embedded platform that needs a custom exit or printf, so it's ok to use
standard functions.

Fixes #8730.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>