pk_wrap: Use Hash algorithm for ECDSA verify_psa
The given hash algorithm for an ecdsa verify operation was just omitted
by setting the algorithm for psa_verify_hash to PSA_ALG_ECDSA_ANY.
As the PSA spec states:
This is the same signature scheme as PSA_ALG_ECDSA(), but without
specifying a hash algorithm, and skipping the message hashing operation.

This algorithm is only recommended to sign or verify a sequence of bytes
that are an already-calculated hash. Note that the input is padded with
zeros on the left or truncated on the right as required to fit the curve
size.

So the input should be hashed but thats not the case for
ecdsa_verify_psa therefore changing it to PSA_ALG_ECDSA(hash_alg)

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>