all.sh: improve components testing CRYPTO_CLIENT
In both cases we check that both libmbedcrypto is build without
missing any mbedtls_xxx() function. Only psa_yyy() are allowed
to be missing due to the lack of PSA provider in the builds.

Note: the component testing the default configuration now is
"downgraded" to only building the library because in the upcoming
commits this would become untestable as well.

Signed-off-by: Valerio Setti <valerio.setti@nordicsemi.no>