Fix TLS 1.2 server certificate filtering with ECDH-only keys
When MBEDTLS_USE_PSA_CRYPTO is disabled and using an opaque ECDH-only key,
don't match the key with an ECDSA certificate.

Also fix an older bug whereby when MBEDTLS_USE_PSA_CRYPTO is enabled, for an
ECC key, the check for a matching curve was incorrectly skipped. This could
have led to an error further in the handshake.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>