Distinguish unmet test function and test suite dependencies
When running a test suite in verbose mode, and in the outcome file, if a
test function is disabled at compile time, indicate whether this is due to a
dependency on the function or on the test suite. This makes it a little
easier to diagnose why a test case is not executed.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>