Use context managers for open files and processes
Using context managers ensures that resources are cleaned up on an
exception. When writing files, they ensure that all data is flushed
out (I think CPython takes care of it anyway, but not Pypy). When reading
files, this only prevents leaking the file descriptor (under Python
implementations that don't close files automatically, such as Pypy), which
doesn't really matter to us since our scripts are short-lived and exit
immediately on error, but it's good practice anyway.

Leave one process call alone in test_psa_compliance because it would take
more refactoring than it's worth. It's in main() so not worth much.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>