Generate certificates with unusual string types
Generate certificates with subject names that have PrintableString,
IA5String and T61String types for testing.

Leave UniversalString BMPString type aside for now. The other string
types behave essentially like various subsets of ASCII for the values
we are using, whereas BMPString is based on UCS-2 (2-byte encoding)
and UniversalString is UTF-32 (4-byte encoding).

It would be good to have tests for these other types but it is out of
scope for a fix to supporting T61 strings.

Signed-off-by: David Horstmann <david.horstmann@arm.com>